+++ resources/views/components/service/farbwahl.blade.php
@@ -26,7 +26,5 @@
             })()">{{ $name ?: __('Vorschau') }}</span>
     </div>
 
-    @error('color')
-        <span class="mt-1 text-sm text-red-600 dark:text-red-400">{{ $message }}</span>
-    @enderror
+    <x-input.fehler feld="color" />
 </div>